pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "honeywell-radio-exporter"
version = "1.0.0"
description = "RAMSES RF Prometheus Exporter"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Honeywell Radio Exporter", email = "simon-honeywellradioexporter@think.io"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
    "Topic :: System :: Monitoring",
]
requires-python = ">=3.11"
dependencies = [
    "prometheus-client>=0.17.0",
    "colorlog>=6.9.0",
    "paho-mqtt>=2.1.0",
    "pyserial-asyncio-fast>=0.16",
    "voluptuous>=0.15.2",
    "asyncio-mqtt>=0.16.0",
    "PyMySQL>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pylint>=2.17.0",
    "flake8>=6.0.0",
    "bandit>=1.7.0",
    "mypy>=1.3.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "safety>=2.3.0",
    "coverage>=7.2.0",
    "tox>=4.0.0",
]

[project.scripts]
honeywell-radio-exporter = "honeywell_radio_exporter.app:main"

[project.urls]
Homepage = "https://github.com/example/honeywell-radio-exporter"
Repository = "https://github.com/example/honeywell-radio-exporter"
Issues = "https://github.com/example/honeywell-radio-exporter/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["honeywell_radio_exporter*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["honeywell_radio_exporter"]
known_third_party = ["prometheus_client", "ramses_rf", "ramses_tx"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["honeywell_radio_exporter"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "*/.tox/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101", "B601"]
